计分系统实现布尔计分
- 布尔计分逻辑实现
- ScoreDataImporter接口实现，引入布尔计分逻辑
- 修改IndexScoreCounter，引入另一参数Context

+++ src/main/java/io/github/packagewjx/brandreportbackend/service/report/score/EnumScoreDefinition.java
@@ -7,13 +7,13 @@ import java.util.Map;
  * @date 19-7-17
  **/
 public class EnumScoreDefinition {
-    private Map<Object, Integer> definition;
+    private Map<Object, Double> definition;
 
-    public Map<Object, Integer> getDefinition() {
+    public Map<Object, Double> getDefinition() {
         return definition;
     }
 
-    public void setDefinition(Map<Object, Integer> definition) {
+    public void setDefinition(Map<Object, Double> definition) {
         this.definition = definition;
     }
 }
